Bump Ruby docker image from 2.7.0 to 3.2.2
Support for Ruby 2.7 was [dropped][1] in bundler 2.5.0, released today
(2023-12-15). We need bundler in this action. Therefore, we're bumping
our Ruby version to a version that bundler supports, 3.2.2. This is also
the most recent stable release of Ruby, and also the version of Ruby
that we are currently using in the `commonlit` repo.

Discussed in Slack [here][2].

[Example][3] of a failing build.

[1]: https://github.com/rubygems/rubygems/blob/master/bundler/CHANGELOG.md#250-december-15-2023
[2]: https://commonlit.slack.com/archives/C5BFTS7NC/p1702662411162479
[3]: https://github.com/commonlit/commonlit/actions/runs/7225260353/job/19688299101?pr=25786#step:2:130